% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{bookdowntargets-package}
\alias{bookdowntargets-package}
\title{bookdowntargets: tarchetype pipeline for bookdown reports.}
\description{
A pipeline toolkit for R, the \code{targets} package brings together
function-oriented programming and Make-like declarative pipelines for
Statistics and data science. The \code{bookdowntargets} package provides a
convenient helper function to create specialized targets.
}
