% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_assert.R
\name{tar_assert_dir}
\alias{tar_assert_dir}
\title{Assertions}
\usage{
tar_assert_dir(x, msg = NULL)
}
\arguments{
\item{x}{R object to be validated.}

\item{msg}{Character of length 1, a message to be printed to the console if
`x` is invalid.}
}
\description{
Asserts the correctness of user inputs and generate custom error conditions
  as needed.
}
\examples{
tar_assert_dir(getwd())
}
\keyword{internal}
